+++ docs/source/release-notes.rst
@@ -9,6 +9,14 @@ The APIs before v3.0.0 are in beta and may change without prior notice.
 v1.0
 ==========
 
+v1.0.10 (2024-xx-xx)
+--------------------
+
+- Add bus type correction in ``system.System.setup()``
+- Revise ``ams.io.psse.read`` to complete model Zone when necessary
+- Use numerical Area and Zone idx in MATPOWER and PSSE RAW file conversion
+- Support JSON format addfile when converting to ANDES case
+
 v1.0.9 (2024-04-23)
 --------------------